moving Provider to standalone file
