+++ Source/Moya.swift
@@ -33,92 +33,6 @@ public protocol Cancellable {
     func cancel()
 }
 
-/// Request provider class. Requests should be made through this class only.
-public class MoyaProvider<Target: TargetType> {
-    
-    /// Closure that defines the endpoints for the provider.
-    public typealias EndpointClosure = Target -> Endpoint<Target>
-    
-    /// Closure that resolves an Endpoint into an NSURLRequest.
-    public typealias RequestClosure = (Endpoint<Target>, NSURLRequest -> Void) -> Void
-    
-    /// Closure that decides if/how a request should be stubbed.
-    public typealias StubClosure = Target -> Moya.StubBehavior
-    
-    public let endpointClosure: EndpointClosure
-    public let requestClosure: RequestClosure
-    public let stubClosure: StubClosure
-    public let manager: Manager
-    
-    /// A list of plugins
-    /// e.g. for logging, network activity indicator or credentials
-    public let plugins: [PluginType]
-    
-    /// Initializes a provider.
-    public init(endpointClosure: EndpointClosure = DefaultEndpointMapping,
-        requestClosure: RequestClosure = DefaultRequestMapping,
-        stubClosure: StubClosure = NeverStub,
-        manager: Manager = DefaultAlamofireManager(),
-        plugins: [PluginType] = []) {
-            
-            self.endpointClosure = endpointClosure
-            self.requestClosure = requestClosure
-            self.stubClosure = stubClosure
-            self.manager = manager
-            self.plugins = plugins
-    }
-    
-    /// Returns an Endpoint based on the token, method, and parameters by invoking the endpointsClosure.
-    public func endpoint(token: Target) -> Endpoint<Target> {
-        return endpointClosure(token)
-    }
-    
-    /// Designated request-making method. Returns a Cancellable token to cancel the request later.
-    public func request(target: Target, completion: Moya.Completion) -> Cancellable {
-        let endpoint = self.endpoint(target)
-        let stubBehavior = self.stubClosure(target)
-        var cancellableToken = CancellableWrapper()
-        
-        let performNetworking = { (request: NSURLRequest) in
-            if cancellableToken.isCancelled { return }
-            
-            switch stubBehavior {
-            case .Never:
-                cancellableToken.innerCancellable = self.sendRequest(target, request: request, completion: completion)
-            default:
-                cancellableToken.innerCancellable = self.stubRequest(target, request: request, completion: completion, endpoint: endpoint, stubBehavior: stubBehavior)
-            }
-        }
-        
-        requestClosure(endpoint, performNetworking)
-        
-        return cancellableToken
-    }
-    
-    /// When overriding this method, take care to `notifyPluginsOfImpendingStub` and to perform the stub using the `createStubFunction` method.
-    /// Note: this was previously in an extension, however it must be in the original class declaration to allow subclasses to override.
-    internal func stubRequest(target: Target, request: NSURLRequest, completion: Moya.Completion, endpoint: Endpoint<Target>, stubBehavior: Moya.StubBehavior) -> CancellableToken {
-        let cancellableToken = CancellableToken { }
-        notifyPluginsOfImpendingStub(request, target: target)
-        let plugins = self.plugins
-        let stub: () -> () = createStubFunction(cancellableToken, forTarget: target, withCompletion: completion, endpoint: endpoint, plugins: plugins)
-        switch stubBehavior {
-        case .Immediate:
-            stub()
-        case .Delayed(let delay):
-            let killTimeOffset = Int64(CDouble(delay) * CDouble(NSEC_PER_SEC))
-            let killTime = dispatch_time(DISPATCH_TIME_NOW, killTimeOffset)
-            dispatch_after(killTime, dispatch_get_main_queue()) {
-                stub()
-            }
-        case .Never:
-            fatalError("Method called to stub request when stubbing is disabled.")
-        }
-        
-        return cancellableToken
-    }
-}
-
 /// Mark: Defaults
 
 // These functions are default mappings to MoyaProvider's properties: endpoints, requests, manager, etc.
@@ -143,9 +57,6 @@ public func DefaultAlamofireManager() -> Manager {
 
 /// Mark: Stubbing
 
-// Swift won't let us put the StubBehavior enum inside the provider class, so we'll
-// at least add some class functions to allow easy access to common stubbing closures.
-
 public func NeverStub<Target: TargetType>(_: Target) -> Moya.StubBehavior {
     return .Never
 }
@@ -156,58 +67,6 @@ public func ImmediatelyStub<Target: TargetType>(_: Target) -> Moya.StubBehavior 
 
 public func DelayedStub<Target: TargetType>(seconds: NSTimeInterval)(_: Target) -> Moya.StubBehavior {
     return .Delayed(seconds: seconds)
-}
-
-internal extension MoyaProvider {
-    
-    func sendRequest(target: Target, request: NSURLRequest, completion: Moya.Completion) -> CancellableToken {
-        let alamoRequest = manager.request(request)
-        let plugins = self.plugins
-        
-        // Give plugins the chance to alter the outgoing request
-        plugins.forEach { $0.willSendRequest(alamoRequest, target: target) }
-        
-        // Perform the actual request
-        alamoRequest.response { (_, response: NSHTTPURLResponse?, data: NSData?, error: NSError?) -> () in
-            let result = convertResponseToResult(response, data: data, error: error)
-            // Inform all plugins about the response
-            plugins.forEach { $0.didReceiveResponse(result, target: target) }
-            completion(result: result)
-        }
-
-        alamoRequest.resume()
-
-        return CancellableToken(request: alamoRequest)
-    }
-    
-    /// Creates a function which, when called, executes the appropriate stubbing behavior for the given parameters.
-    internal final func createStubFunction(token: CancellableToken, forTarget target: Target, withCompletion completion: Moya.Completion, endpoint: Endpoint<Target>, plugins: [PluginType]) -> (() -> ()) {
-        return {
-            if (token.canceled) {
-                let error = Moya.Error.Underlying(NSError(domain: NSURLErrorDomain, code: NSURLErrorCancelled, userInfo: nil))
-                plugins.forEach { $0.didReceiveResponse(.Failure(error), target: target) }
-                completion(result: .Failure(error))
-                return
-            }
-            
-            switch endpoint.sampleResponseClosure() {
-            case .NetworkResponse(let statusCode, let data):
-                let response = Moya.Response(statusCode: statusCode, data: data, response: nil)
-                plugins.forEach { $0.didReceiveResponse(.Success(response), target: target) }
-                completion(result: .Success(response))
-            case .NetworkError(let error):
-                let error = Moya.Error.Underlying(error)
-                plugins.forEach { $0.didReceiveResponse(.Failure(error), target: target) }
-                completion(result: .Failure(error))
-            }
-        }
-    }
-    
-    /// Notify all plugins that a stub is about to be performed. You must call this if overriding `stubRequest`.
-    internal final func notifyPluginsOfImpendingStub(request: NSURLRequest, target: Target) {
-        let alamoRequest = manager.request(request)
-        plugins.forEach { $0.willSendRequest(alamoRequest, target: target) }
-    }
 }
 
 internal func convertResponseToResult(response: NSHTTPURLResponse?, data: NSData?, error: NSError?) ->
@@ -222,15 +81,5 @@ internal func convertResponseToResult(response: NSHTTPURLResponse?, data: NSData
     default:
         let error = Moya.Error.Underlying(NSError(domain: NSURLErrorDomain, code: NSURLErrorUnknown, userInfo: nil))
         return .Failure(error)
-    }
-}
-
-private struct CancellableWrapper: Cancellable {
-    var innerCancellable: CancellableToken? = nil
-    
-    private var isCancelled = false
-    
-    func cancel() {
-        innerCancellable?.cancel()
     }
 }
