+++ Source/Provider.swift
@@ -1,0 +1,149 @@
+import Foundation
+
+/// Request provider class. Requests should be made through this class only.
+public class MoyaProvider<Target: TargetType> {
+
+    /// Closure that defines the endpoints for the provider.
+    public typealias EndpointClosure = Target -> Endpoint<Target>
+
+    /// Closure that resolves an Endpoint into an NSURLRequest.
+    public typealias RequestClosure = (Endpoint<Target>, NSURLRequest -> Void) -> Void
+
+    /// Closure that decides if/how a request should be stubbed.
+    public typealias StubClosure = Target -> Moya.StubBehavior
+
+    public let endpointClosure: EndpointClosure
+    public let requestClosure: RequestClosure
+    public let stubClosure: StubClosure
+    public let manager: Manager
+
+    /// A list of plugins
+    /// e.g. for logging, network activity indicator or credentials
+    public let plugins: [PluginType]
+
+    /// Initializes a provider.
+    public init(endpointClosure: EndpointClosure = DefaultEndpointMapping,
+                requestClosure: RequestClosure = DefaultRequestMapping,
+                stubClosure: StubClosure = NeverStub,
+                manager: Manager = DefaultAlamofireManager(),
+                plugins: [PluginType] = []) {
+
+        self.endpointClosure = endpointClosure
+        self.requestClosure = requestClosure
+        self.stubClosure = stubClosure
+        self.manager = manager
+        self.plugins = plugins
+    }
+
+    /// Returns an Endpoint based on the token, method, and parameters by invoking the endpointsClosure.
+    public func endpoint(token: Target) -> Endpoint<Target> {
+        return endpointClosure(token)
+    }
+
+    /// Designated request-making method. Returns a Cancellable token to cancel the request later.
+    public func request(target: Target, completion: Moya.Completion) -> Cancellable {
+        let endpoint = self.endpoint(target)
+        var cancellableToken = CancellableWrapper()
+        let stubBehavior = self.stubClosure(target)
+
+        let performNetworking = { (request: NSURLRequest) in
+            if cancellableToken.isCancelled { return }
+
+            switch stubBehavior {
+            case .Never:
+                cancellableToken.innerCancellable = self.sendRequest(target, request: request, completion: completion)
+            default:
+                cancellableToken.innerCancellable = self.stubRequest(target, request: request, completion: completion, endpoint: endpoint, stubBehavior: stubBehavior)
+            }
+        }
+
+        requestClosure(endpoint, performNetworking)
+
+        return cancellableToken
+    }
+
+    /// When overriding this method, take care to `notifyPluginsOfImpendingStub` and to perform the stub using the `createStubFunction` method.
+    /// Note: this was previously in an extension, however it must be in the original class declaration to allow subclasses to override.
+    internal func stubRequest(target: Target, request: NSURLRequest, completion: Moya.Completion, endpoint: Endpoint<Target>, stubBehavior: Moya.StubBehavior) -> CancellableToken {
+        let cancellableToken = CancellableToken { }
+        notifyPluginsOfImpendingStub(request, target: target)
+        let plugins = self.plugins
+        let stub: () -> () = createStubFunction(cancellableToken, forTarget: target, withCompletion: completion, endpoint: endpoint, plugins: plugins)
+        switch stubBehavior {
+        case .Immediate:
+            stub()
+        case .Delayed(let delay):
+            let killTimeOffset = Int64(CDouble(delay) * CDouble(NSEC_PER_SEC))
+            let killTime = dispatch_time(DISPATCH_TIME_NOW, killTimeOffset)
+            dispatch_after(killTime, dispatch_get_main_queue()) {
+                stub()
+            }
+        case .Never:
+            fatalError("Method called to stub request when stubbing is disabled.")
+        }
+
+        return cancellableToken
+    }
+}
+
+internal extension MoyaProvider {
+
+    func sendRequest(target: Target, request: NSURLRequest, completion: Moya.Completion) -> CancellableToken {
+        let alamoRequest = manager.request(request)
+        let plugins = self.plugins
+
+        // Give plugins the chance to alter the outgoing request
+        plugins.forEach { $0.willSendRequest(alamoRequest, target: target) }
+
+        // Perform the actual request
+        alamoRequest.response { (_, response: NSHTTPURLResponse?, data: NSData?, error: NSError?) -> () in
+            let result = convertResponseToResult(response, data: data, error: error)
+            // Inform all plugins about the response
+            plugins.forEach { $0.didReceiveResponse(result, target: target) }
+            completion(result: result)
+        }
+
+        alamoRequest.resume()
+
+        return CancellableToken(request: alamoRequest)
+    }
+
+    /// Creates a function which, when called, executes the appropriate stubbing behavior for the given parameters.
+    internal final func createStubFunction(token: CancellableToken, forTarget target: Target, withCompletion completion: Moya.Completion, endpoint: Endpoint<Target>, plugins: [PluginType]) -> (() -> ()) {
+        return {
+            if (token.canceled) {
+                let error = Moya.Error.Underlying(NSError(domain: NSURLErrorDomain, code: NSURLErrorCancelled, userInfo: nil))
+                plugins.forEach { $0.didReceiveResponse(.Failure(error), target: target) }
+                completion(result: .Failure(error))
+                return
+            }
+
+            switch endpoint.sampleResponseClosure() {
+            case .NetworkResponse(let statusCode, let data):
+                let response = Moya.Response(statusCode: statusCode, data: data, response: nil)
+                plugins.forEach { $0.didReceiveResponse(.Success(response), target: target) }
+                completion(result: .Success(response))
+            case .NetworkError(let error):
+                let error = Moya.Error.Underlying(error)
+                plugins.forEach { $0.didReceiveResponse(.Failure(error), target: target) }
+                completion(result: .Failure(error))
+            }
+        }
+    }
+
+    /// Notify all plugins that a stub is about to be performed. You must call this if overriding `stubRequest`.
+    internal final func notifyPluginsOfImpendingStub(request: NSURLRequest, target: Target) {
+        let alamoRequest = manager.request(request)
+        plugins.forEach { $0.willSendRequest(alamoRequest, target: target) }
+    }
+}
+
+private struct CancellableWrapper: Cancellable {
+    var innerCancellable: CancellableToken? = nil
+
+    private var isCancelled = false
+
+    func cancel() {
+        innerCancellable?.cancel()
+    }
+}
